delete some unused codes and organize components inside daily-event-card into individual sub folder
