+++ my-app/src/components/DailyEventCard/DayCalendar.js
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from 'react';
 import { Calendar, momentLocalizer, Views } from 'react-big-calendar';
 import moment from 'moment';
 import 'react-big-calendar/lib/addons/dragAndDrop/styles.css';
-import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
 const DateTitle = styled.div`
@@ -10,7 +9,6 @@ const DateTitle = styled.div`
   font-size: 14px;
 `;
 export default function DayCalendar(props) {
-  //   const [dayTimeBlocks, setDayTimeBlocks] = useState([]);
   const localizer = momentLocalizer(moment);
 
   return (
